requirements.txt
aiohttp==3.6.2
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.9.1
bs4==0.0.1
chardet==3.0.4
dateparser==0.7.1
discord==1.0.1
discord.py==1.3.3
future==0.17.1
html2text==2020.1.16
idna==2.8
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.5.1
mccabe==0.6.1
multidict==4.5.2
parsedatetime==2.6
Pillow==7.1.2
PoE.py==2.0.6a0
pylint==2.3.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
regex==2020.6.8
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
typing-extensions==3.7.4
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.3
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
youtube-dl==2020.6.16.1
